SQUASH: sample data
