--- q2_types/_sample_data.py
@@ -1,63 +1,0 @@
-# ----------------------------------------------------------------------------
-# Copyright (c) 2016--, QIIME development team.
-#
-# Distributed under the terms of the Modified BSD License.
-#
-# The full license is in the file LICENSE, distributed with this software.
-# ----------------------------------------------------------------------------
-
-import pandas as pd
-from qiime.plugin import SemanticType
-import qiime.plugin.model as model
-
-from .plugin_setup import plugin
-
-
-SampleData = SemanticType('SampleData', field_names='type')
-
-AlphaDiversity = SemanticType('AlphaDiversity',
-                              variant_of=SampleData.field['type'])
-
-
-# Formats
-class AlphaDiversityFormat(model.TextFileFormat):
-    def sniff(self):
-        with self.open() as fh:
-            for line, _ in zip(fh, range(10)):
-                cells = line.split('\t')
-                if len(cells) != 2:
-                    return False
-            return True
-
-
-AlphaDiversityDirectoryFormat = model.SingleFileDirectoryFormat(
-    'AlphaDiversityDirectoryFormat', 'alpha-diversity.tsv',
-    AlphaDiversityFormat)
-
-
-# Transformers
-@plugin.register_transformer
-def _1(data: pd.Series) -> AlphaDiversityFormat:
-    ff = AlphaDiversityFormat()
-    with ff.open() as fh:
-        data.to_csv(fh, sep='\t', header=True)
-    return ff
-
-
-@plugin.register_transformer
-def _2(ff: AlphaDiversityFormat) -> pd.Series:
-    with ff.open() as fh:
-        # Since we're wanting to round-trip with pd.Series.to_csv, the pandas
-        # docs recommend using from_csv here (rather than the more commonly
-        # used pd.read_csv).
-        return pd.Series.from_csv(fh, sep='\t', header=0)
-
-
-# Registrations
-plugin.register_semantic_type(SampleData)
-plugin.register_semantic_type(AlphaDiversity)
-
-plugin.register_semantic_type_to_format(
-    SampleData[AlphaDiversity],
-    artifact_format=AlphaDiversityDirectoryFormat
-)
