+++ q2_types/plugin_setup.py
@@ -32,7 +32,7 @@ importlib.import_module('q2_types.feature_table')
 importlib.import_module('q2_types.distance_matrix')
 importlib.import_module('q2_types.tree')
 importlib.import_module('q2_types.ordination')
-importlib.import_module('q2_types._sample_data')
+importlib.import_module('q2_types.sample_data')
 importlib.import_module('q2_types.feature_data')
 importlib.import_module('q2_types.reference_features')
 importlib.import_module('q2_types.per_sample_sequences')
