+++ q2_types/sample_data/__init__.py
@@ -1,0 +1,17 @@
+# ----------------------------------------------------------------------------
+# Copyright (c) 2016--, QIIME development team.
+#
+# Distributed under the terms of the Modified BSD License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# ----------------------------------------------------------------------------
+
+import importlib
+
+from ._format import AlphaDiversityFormat, AlphaDiversityDirectoryFormat
+from ._type import SampleData, AlphaDiversity
+
+__all__ = ['AlphaDiversityFormat', 'AlphaDiversityDirectoryFormat',
+           'SampleData', 'AlphaDiversity']
+
+importlib.import_module('q2_types.sample_data._transformer')
